refactor(isolate): rename dispatcher to handler.
