+++ lib/nacht_sources.dart
@@ -13,8 +13,8 @@ export 'src/impls/impls.dart';
 export 'src/interfaces/interfaces.dart';
 
 // Isolate.
-export 'src/isolate/isolate_dispatcher.dart';
 export 'src/isolate/isolate_handler.dart';
+export 'src/isolate/isolate_worker.dart';
 
 // Models.
 export 'src/models/models.dart';
